+++ infrastructure/migrations/versions/001_create_questions.py
@@ -23,11 +23,11 @@ def upgrade() -> None:
     op.create_table(
         "questions",
         sa.Column("token", sa.String(length=255), nullable=False),
-        sa.Column("group_id", sa.BIGINT(), nullable=False),
+        sa.Column("group_id", sa.BigInteger(), nullable=False),
         sa.Column("topic_id", sa.Integer(), nullable=False),
         sa.Column("topic_duty_fullname", sa.Unicode(), nullable=True),
         sa.Column("employee_fullname", sa.Unicode(), nullable=False),
-        sa.Column("employee_chat_id", sa.BIGINT(), nullable=False),
+        sa.Column("employee_chat_id", sa.BigInteger(), nullable=False),
         sa.Column("employee_division", sa.Unicode(), nullable=True),
         sa.Column("question_text", sa.Unicode(), nullable=True),
         sa.Column("start_time", sa.DateTime(), nullable=True),
@@ -40,7 +40,7 @@ def upgrade() -> None:
         sa.Column("activity_status_enabled", sa.Boolean(), nullable=True, default=None),
         sa.PrimaryKeyConstraint("token"),
     )
-    # ### end Alembic commands ###
+    # ### end Alembic commands ###4
 
 
 def downgrade() -> None:
